Updates to footer links to about us page
Links to files have been added to each footer link in all pages.  Also,
on about us page, picture of guitar neck been rescaled so it's not too
big, but still covers as a banner.  I'm not quite sure how to bring
browser to a certain page and to a certain point, got it working in
about us page, but no where else.  Any ideas?